# clients/go/api/openapi.yaml
openapi: 3.0.0
info:
  description: "API REST gamificada para gerenciamento de tarefas com IA, XP, levels\
    \ e sistema de recompensas"
  title: Gamified Task Management API
  version: 1.0.0
servers:
- description: Servidor de Desenvolvimento
  url: http://localhost:3000
security:
- bearerAuth: []
...
paths:
...
  /api/tasks:
    get:
      operationId: listTasks
      parameters:
      - description: Filtrar por status
        explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - pending
          - in_progress
          - completed
          type: string
        style: form
      - description: Filtrar por prioridade
        explode: true
        in: query
        name: priority
        required: false
        schema:
          enum:
          - low
          - medium
          - high
          type: string
        style: form
      - description: Número da página
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: Itens por página
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - description: Campo para ordenação
        explode: true
        in: query
        name: sortBy
        required: false
        schema:
          default: createdAt
          enum:
          - createdAt
          - updatedAt
          - dueDate
          - priority
          - title
          type: string
        style: form
      - description: Ordem da classificação
        explode: true
        in: query
        name: sortOrder
        required: false
        schema:
          default: desc
          enum:
          - asc
          - desc
          type: string
        style: form
      - description: Buscar no título e descrição
        explode: true
        in: query
        name: search
        required: false
        schema:
          maxLength: 100
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listTasks_200_response'
          description: Lista de tarefas
      security:
      - bearerAuth: []
      summary: Get all tasks for the logged-in user
      tags:
      - Tasks
    post:
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createTask_request'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createTask_201_response'
          description: Tarefa criada com sucesso
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Dados inválidos
      security:
      - bearerAuth: []
      summary: Create a new task
      tags:
      - Tasks
  /api/tasks/stats:
    get:
      operationId: getTaskStats
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTaskStats_200_response'
          description: Estatísticas das tarefas
      security:
      - bearerAuth: []
      summary: Get task statistics
      tags:
      - Tasks
...
  /api/tasks/{id}:
    delete:
      operationId: deleteTask
      parameters:
      - description: ID da tarefa
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/deleteTask_200_response'
          description: Tarefa deletada com sucesso
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa não encontrada
      security:
      - bearerAuth: []
      summary: Delete a task
      tags:
      - Tasks
    get:
      operationId: getTaskById
      parameters:
      - description: ID da tarefa
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTaskById_200_response'
          description: Detalhes da tarefa
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa não encontrada
      security:
      - bearerAuth: []
      summary: Get a specific task by ID
      tags:
      - Tasks
    put:
      operationId: updateTask
      parameters:
      - description: ID da tarefa
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateTask_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/updateTask_200_response'
          description: Tarefa atualizada com sucesso
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Dados inválidos
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa não encontrada
      security:
      - bearerAuth: []
      summary: Update a task
      tags:
      - Tasks
  /api/tasks/{id}/complete:
    patch:
      operationId: completeTask
      parameters:
      - description: ID da tarefa
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/completeTask_200_response'
          description: Tarefa marcada como concluída
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa já está concluída
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa não encontrada
      security:
      - bearerAuth: []
      summary: Mark a task as complete
      tags:
      - Tasks
...
  /api/tasks/{id}/complete-quest:
    post:
      operationId: completeQuest
      parameters:
      - description: ID da tarefa/quest
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/completeQuest_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/completeQuest_200_response'
          description: Quest completada com sucesso
      security:
      - bearerAuth: []
      summary: Complete a quest (task) and get XP
      tags:
      - Gamification
...
  /api/users/stats:
    get:
      operationId: getUserStats
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getUserStats_200_response'
          description: Estatísticas do sistema
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado - privilégios de admin necessários
      summary: Obter estatísticas gerais do sistema (Admin only)
      tags:
      - Users
...
components:
  schemas:
...
    Error:
      example:
        success: false
        message: Erro de validação
      properties:
        success:
          example: false
          type: boolean
        message:
          example: Erro de validação
          type: string
      type: object
...
    listTasks_200_response_data:
      example:
        pagination:
          current: 1
          total: 50
          pages: 5
          limit: 10
        tasks:
        - id: 64a1b2c3d4e5f6789abcdef1
          title: Implementar Story-Mode
          priority: low
          userId: 64a1b2c3d4e5f6789abcdef0
          status: pending
        - id: 64a1b2c3d4e5f6789abcdef1
          title: Implementar Story-Mode
          priority: low
          userId: 64a1b2c3d4e5f6789abcdef0
          status: pending
      properties:
        tasks:
          items:
            $ref: '#/components/schemas/Task'
          type: array
        pagination:
          $ref: '#/components/schemas/listTasks_200_response_data_pagination'
      type: object
    listTasks_200_response:
      example:
        data:
          pagination:
            current: 1
            total: 50
            pages: 5
            limit: 10
          tasks:
          - id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
          - id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
        success: true
      properties:
        success:
          example: true
          type: boolean
        data:
          $ref: '#/components/schemas/listTasks_200_response_data'
      type: object
    createTask_request:
      properties:
        title:
          example: Implementar autenticação JWT
          maxLength: 100
          minLength: 3
          type: string
        description:
          example: Implementar sistema de autenticação usando JWT com middleware de
            verificação
          maxLength: 500
          minLength: 10
          type: string
        status:
          default: pending
          enum:
          - pending
          - in_progress
          - completed
          example: pending
          type: string
        priority:
          default: medium
          enum:
          - low
          - medium
          - high
          example: high
          type: string
        dueDate:
          example: 2024-12-31T23:59:59Z
          format: date-time
          type: string
        tags:
          example:
          - backend
          - auth
          - jwt
          items:
            maxLength: 20
            type: string
          maxItems: 10
          type: array
      required:
      - description
      - title
      type: object
    createTask_201_response_data:
      example:
        task:
          id: 64a1b2c3d4e5f6789abcdef1
          title: Implementar Story-Mode
          priority: low
          userId: 64a1b2c3d4e5f6789abcdef0
          status: pending
      properties:
        task:
          $ref: '#/components/schemas/Task'
      type: object
    createTask_201_response:
      example:
        data:
          task:
            id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
        success: true
        message: Tarefa criada com sucesso
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Tarefa criada com sucesso
          type: string
        data:
          $ref: '#/components/schemas/createTask_201_response_data'
      type: object
...
    getTaskStats_200_response_data:
      example:
        stats:
          total: 25
          in_progress: 8
          overdue: 3
          completedThisMonth: 5
          pending: 10
          completed: 7
        overdueTasks:
        - id: 64a1b2c3d4e5f6789abcdef1
          title: Implementar Story-Mode
          priority: low
          userId: 64a1b2c3d4e5f6789abcdef0
          status: pending
        - id: 64a1b2c3d4e5f6789abcdef1
          title: Implementar Story-Mode
          priority: low
          userId: 64a1b2c3d4e5f6789abcdef0
          status: pending
      properties:
        stats:
          $ref: '#/components/schemas/getTaskStats_200_response_data_stats'
        overdueTasks:
          items:
            $ref: '#/components/schemas/Task'
          type: array
      type: object
    getTaskStats_200_response:
      example:
        data:
          stats:
            total: 25
            in_progress: 8
            overdue: 3
            completedThisMonth: 5
            pending: 10
            completed: 7
          overdueTasks:
          - id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
          - id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
        success: true
      properties:
        success:
          example: true
          type: boolean
        data:
          $ref: '#/components/schemas/getTaskStats_200_response_data'
      type: object
...
    getTaskById_200_response:
      example:
        data:
          task:
            id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
        success: true
      properties:
        success:
          example: true
          type: boolean
        data:
          $ref: '#/components/schemas/createTask_201_response_data'
      type: object
    updateTask_request:
      minProperties: 1
      properties:
        title:
          example: Implementar autenticação JWT - Atualizado
          maxLength: 100
          minLength: 3
          type: string
        description:
          example: Implementar sistema de autenticação usando JWT com middleware de
            verificação e refresh tokens
          maxLength: 500
          minLength: 10
          type: string
        status:
          enum:
          - pending
          - in_progress
          - completed
          example: in_progress
          type: string
        priority:
          enum:
          - low
          - medium
          - high
          example: high
          type: string
        dueDate:
          example: 2024-12-31T23:59:59Z
          format: date-time
          nullable: true
          type: string
        tags:
          example:
          - backend
          - auth
          - jwt
          - security
          items:
            maxLength: 20
            type: string
          maxItems: 10
          type: array
      type: object
    updateTask_200_response:
      example:
        data:
          task:
            id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
        success: true
        message: Tarefa atualizada com sucesso
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Tarefa atualizada com sucesso
          type: string
        data:
          $ref: '#/components/schemas/createTask_201_response_data'
      type: object
    deleteTask_200_response:
      example:
        success: true
        message: Tarefa deletada com sucesso
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Tarefa deletada com sucesso
          type: string
      type: object
    completeTask_200_response:
      example:
        data:
          task:
            id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
        success: true
        message: Tarefa marcada como concluída
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Tarefa marcada como concluída
          type: string
        data:
          $ref: '#/components/schemas/createTask_201_response_data'
      type: object
...
    completeQuest_request:
      properties:
        completionTime:
          description: Tempo gasto em minutos
          example: 45
          type: number
      type: object
...
    completeQuest_200_response_data:
      example:
        gamification:
          xpGained: 6.027456183070403
          leveledUp: true
          totalXP: 5.962133916683182
          currentStreak: 5.637376656633329
          newLevel: 1.4658129805029452
          newBadges:
          - newBadges
          - newBadges
        task:
          id: id
          title: title
          questTitle: questTitle
          xpEarned: 0.8008281904610115
      properties:
        task:
          $ref: '#/components/schemas/completeQuest_200_response_data_task'
        gamification:
          $ref: '#/components/schemas/completeQuest_200_response_data_gamification'
      type: object
    completeQuest_200_response:
      example:
        data:
          gamification:
            xpGained: 6.027456183070403
            leveledUp: true
            totalXP: 5.962133916683182
            currentStreak: 5.637376656633329
            newLevel: 1.4658129805029452
            newBadges:
            - newBadges
            - newBadges
          task:
            id: id
            title: title
            questTitle: questTitle
            xpEarned: 0.8008281904610115
        success: true
        message: Quest completada! +25 XP
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Quest completada! +25 XP
          type: string
        data:
          $ref: '#/components/schemas/completeQuest_200_response_data'
      type: object
...
    getUserStats_200_response:
      example:
        data:
          mostActiveUsers:
          - taskCount: 0
            name: name
            userId: userId
            email: email
          - taskCount: 0
            name: name
            userId: userId
            email: email
          users:
            total: 150
            inactive: 10
            active: 140
            admins: 5
          tasks:
            total: 1250
            inProgress: 150
            overdue: 45
            pending: 300
            completed: 800
        success: true
      properties:
        success:
          example: true
          type: boolean
        data:
          $ref: '#/components/schemas/getUserStats_200_response_data'
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http